requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# streamdemo/setup.py: 11
names == 0.3.0

# streamdemo/setup.py: 12
networkx == 1.11

# streamdemo/algo.py: 8
# streamdemo/tests/test_algo.py: 1
numpy == 1.13.1

# streamdemo/__init__.py: 36
# streamdemo/feedtools.py: 3
pymongo == 3.4.0

# rundemo.py: 10
# streamdemo/setup.py: 13
stream_python == 2.3.11
